+++ src/main/java/io/github/jamalam360/jamlib/Callback.java
@@ -1,0 +1,10 @@
+package io.github.jamalam360.jamlib;
+
+/**
+ * @author Jamalam
+ */
+
+@FunctionalInterface
+public interface Callback {
+    void accept();
+}