+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Changelog
 
-Fixes a reflection issue with the registry module.
+### Features
+
+- Allow `SoundEvent`s, `RecipeSerializer`s, and `RecipeType`s to be registered.
+- Add a `Util` class.
 
 Closed issues: None.
 
-[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.4.2...0.4.3)
+[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.4.3...0.4.4)